utils/rl_autoscheduler/requirements.txt
PyYAML == 6.0
gym == 0.21.0
numpy == 1.23.1
ray[rllib] == 1.13.0
sympy == 1.10.1
torch == 1.12.0
tqdm == 4.64.0
